Start conversion to accepting CSV files and support for custom fields
It turns out EPPlus isn't really .Net Core friedly as under the hood it uses OLE.
As a result, the simplest option is to support CSV for now and do our mapping from
columns/rows to a Dictionary<String,String> as early as possible.

Later it will be possible to use a dependency free .xlsx reader to handle Excel
files, but for the moment I'm more concerned with getting my import done.

Additionally this commit removes some archaic/non-best practices and changes
the field names to better match what they refer to. I.e. name -> email because
we're logging in by email.