Implement support for custom fields. Both simplify and complicate
What's new:

1. Support for custom fields.

Now the header values are what are sent to Fogbugz.

2. Refactor to separate concers.

The code that talks to FogBugz and the code that chews up CSVs are separate.
This will make it possible to write tests without doing crazy things.

3. Async all the things

4. Start parallelizing file processing

5. Add an actual command line option parser

6. Cleanup debugging

Mostly this means removing the existing ad hoc debugging.
Additionally I added code/logging/messaging that will throw an error earlier in the
process when things are not setup correctly.

Next up is adding serilog or something else sufficiently easy to integrate so we can
log:

1. What are we sending to fogbugz (to do a dry run, for example)
2. Verbose error logging. Because who doesn't enjoy stack traces right in the middle
of your console output.
3. At different levels of verbosity. Maybe you want an error message but not a stack
trace. Maybe you want to know everything.